FEATURE: Move StandardsComplianceComponent to Middleware
